chore(organize folder structure): clean up. file and folder configuration
