+++ components/SlateEditor/Editor/Settings/editableProps.ts
