+++ components/SlateEditor/Editor/Settings/plateTypes.ts
