+++ components/SlateEditor/customPlugins/Linebreaks/softBreakPlugin.tsx
@@ -4,7 +4,7 @@ import {
   ELEMENT_TD,
   SoftBreakPlugin,
 } from '@udecode/plate';
-import { MyPlatePlugin } from '../plateTypes';
+import { MyPlatePlugin } from '../../Editor/Settings/plateTypes';
 
 export const softBreakPlugin: Partial<MyPlatePlugin<SoftBreakPlugin>> = {
   options: {
