+++ components/SlateEditor/customPlugins/Link/linkPlugin.tsx
@@ -3,7 +3,7 @@ import {
   PlateFloatingLink,
   RenderAfterEditable,
 } from '@udecode/plate';
-import { MyPlatePlugin, MyValue } from '../plateTypes';
+import { MyPlatePlugin, MyValue } from '../../Editor/Settings/plateTypes';
 
 export const linkPlugin: Partial<MyPlatePlugin<LinkPlugin>> = {
   renderAfterEditable: PlateFloatingLink as RenderAfterEditable<MyValue>,
